+++ MintedEthCappedCrowdsaleCombined.sol
@@ -440,7 +440,7 @@ contract FinalizeAgent {
   function isSane(address crowdsale) public constant returns (bool) {
     return true;
   }
-  function calculatePrice(uint value, uint weiRaised, uint tokensSold, address msgSender, uint8 decimals) public constant returns (uint tokenAmount);
+  function calculatePrice(uint value, uint weiRaised, uint tokensSold, address msgSender, uint decimals) public constant returns (uint tokenAmount);
 }
 
 
@@ -564,7 +564,7 @@ contract Crowdsale is Haltable, SafeMathLib {
   }
   function preallocate(address receiver, uint fullTokens, uint weiPrice) public onlyOwner {
 
-    uint tokenAmount = fullTokens * 10**token.decimals();
+    uint tokenAmount = fullTokens * 10**uint(token.decimals());
     uint weiAmount = weiPrice * fullTokens;
 
     weiRaised = safeAdd(weiRaised,weiAmount);
