+++ contracts/Crowdsale.sol
@@ -230,7 +230,7 @@ contract Crowdsale is Haltable, SafeMathLib {
    */
   function preallocate(address receiver, uint fullTokens, uint weiPrice) public onlyOwner {
 
-    uint tokenAmount = fullTokens * 10**token.decimals();
+    uint tokenAmount = fullTokens * 10**uint(token.decimals());
     uint weiAmount = weiPrice * fullTokens; // This can be also 0, we give out tokens for free
 
     weiRaised = safeAdd(weiRaised,weiAmount);
