+++ contracts/TokenTranchePricing.sol
@@ -128,7 +128,7 @@ contract TokenTranchePricing is PricingStrategy, Ownable, SafeMathLib {
   }
 
   /// @dev Calculate the current price for buy in amount.
-  function calculatePrice(uint value, uint weiRaised, uint tokensSold, address msgSender, uint8 decimals) public constant returns (uint) {
+  function calculatePrice(uint value, uint weiRaised, uint tokensSold, address msgSender, uint decimals) public constant returns (uint) {
 
     uint multiplier = 10 ** decimals;
